requirements.txt
aiogram>=3.0
loguru
python-dotenv
tortoise-orm
aiosqlite
aerich
tomlkit
aiohttp
apscheduler>=3.10,<4.0
